updated Wes corrections
