+++ .gitignore
@@ -8,3 +8,6 @@ electricitylci/__pycache__/elci_database_generator.cpython-36.pyc
 *.pyc
 *.xlsx
 *.xlsx
+electricitylci/__pycache__/elci_database_generator.cpython-36.pyc
+electricitylci/save.p
+*.pyc
